fix remaining devtools commands
