+++ rats-devtools/src/python/rats/devtools/__init__.py
@@ -1,10 +1,8 @@
 """Some utilities for developing the devtool cli commands."""
 
 from ._app import AppServices, run
-from ._plugin import PluginServices
 
 __all__ = [
-    "PluginServices",
     "AppServices",
     "run",
 ]
