+++ rats-devtools/src/python/rats/docs/_plugin.py
@@ -21,11 +21,11 @@ class PluginServices:
 
 class PluginContainer(apps.Container, apps.PluginMixin):
 
-    @apps.group(devtools.PluginServices.EVENTS.OPENING)
+    @apps.group(devtools.AppServices.ON_REGISTER)
     def _on_open(self) -> Iterator[apps.Executable]:
         yield apps.App(
             lambda: cli.attach(
-                self._app.get(devtools.PluginServices.MAIN_CLICK),
+                self._app.get(devtools.AppServices.MAIN_CLICK),
                 self._app.get(PluginServices.MAIN_CLICK),
             )
         )
